pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Voice-first local coding assistant. vLLM backend, faster-whisper STT, Piper TTS, Telegram bridge."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Rui Machado" }]

dependencies = [
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "faster-whisper>=1.0.3",
    "sounddevice>=0.4.7",
    "numpy>=1.26.0",
    "webrtcvad>=2.0.10",
    "piper-tts>=1.2.0; platform_system != 'Darwin'",
    "python-telegram-bot>=21.4",
    "anyio>=4.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "ruff>=0.5.0",
]

[project.scripts]
jarvis = "jarvis.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jarvis"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
